Time: 1 ms (100.00%) | Memory: 42.4 MB (82.46%) - LeetSync

+++ 3626-smallest-divisible-digit-product-i/smallest-divisible-digit-product-i.java
@@ -4,7 +4,7 @@ class Solution {
         
         while(true)
         {
-            long p=1;
+            int p=1;
             int temp=n;
 
             while(temp>0)
